Allow render_docs.py tool to be run on any directory
